+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/EventDetailPresenterTestBuilder.swift
@@ -4,41 +4,6 @@ import EurofurenceModelTestDoubles
 import Foundation
 import UIKit.UIViewController
 
-class CapturingInteraction: Interaction {
-    
-    enum State {
-        case unset
-        case active
-        case inactive
-    }
-    
-    private(set) var state: State = .unset
-    
-    func activate() {
-        state = .active
-    }
-    
-    func deactivate() {
-        state = .inactive
-    }
-    
-}
-
-class CapturingEventInteractionRecorder: EventInteractionRecorder {
-    
-    private(set) var witnessedEvent: EventIdentifier?
-    private(set) var currentInteraction: CapturingInteraction?
-    func makeInteraction(for event: EventIdentifier) -> Interaction? {
-        witnessedEvent = event
-        
-        let interaction = CapturingInteraction()
-        currentInteraction = interaction
-        
-        return interaction
-    }
-    
-}
-
 class EventDetailPresenterTestBuilder {
 
     struct Context {
