+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/CapturingInteraction.swift
@@ -1,0 +1,21 @@
+import Eurofurence
+
+class CapturingInteraction: Interaction {
+    
+    enum State {
+        case unset
+        case active
+        case inactive
+    }
+    
+    private(set) var state: State = .unset
+    
+    func activate() {
+        state = .active
+    }
+    
+    func deactivate() {
+        state = .inactive
+    }
+    
+}
