+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/DummyEventDetailViewModelFactory.swift
