+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubActionEventViewModel.swift
@@ -1,0 +1,14 @@
+import Eurofurence
+
+struct StubActionEventViewModel: EventDetailViewModel {
+    
+    var actionViewModel: EventActionViewModel
+    
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(actionViewModel
+        ) }
+    func favourite() { }
+    func unfavourite() { }
+    
+}
