+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubArtShowEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubArtShowEventViewModel: EventDetailViewModel {
+
+    var artShowMessageViewModel: EventArtShowMessageViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(artShowMessageViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
