+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubDealersDenEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubDealersDenEventViewModel: EventDetailViewModel {
+
+    var dealersDenMessageViewModel: EventDealersDenMessageViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(dealersDenMessageViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
