+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubEventDetailSceneFactory.swift
