+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubKageEventViewModel.swift
@@ -1,0 +1,14 @@
+import Eurofurence
+
+struct StubKageEventViewModel: EventDetailViewModel {
+
+    var kageMessageViewModel: EventKageMessageViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(kageMessageViewModel
+        ) }
+    func favourite() { }
+    func unfavourite() { }
+
+}
