+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubMainStageEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubMainStageEventViewModel: EventDetailViewModel {
+
+    var mainStageMessageViewModel: EventMainStageMessageViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(mainStageMessageViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
