+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubPhotoshootEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubPhotoshootEventViewModel: EventDetailViewModel {
+
+    var photoshootMessageViewModel: EventPhotoshootMessageViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(photoshootMessageViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
