+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubSponsorsOnlyEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubSponsorsOnlyEventViewModel: EventDetailViewModel {
+
+    var sponsorsOnlyWarningViewModel: EventSponsorsOnlyWarningViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(sponsorsOnlyWarningViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
