+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/Test Doubles/View Models/StubSuperSponsorsOnlyEventViewModel.swift
@@ -1,0 +1,16 @@
+import Eurofurence
+
+struct StubSuperSponsorsOnlyEventViewModel: EventDetailViewModel {
+
+    var superSponsorsOnlyWarningViewModel: EventSuperSponsorsOnlyWarningViewModel
+
+    var numberOfComponents: Int { return 1 }
+    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
+    
+    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(superSponsorsOnlyWarningViewModel)
+    }
+    
+    func favourite() { }
+    func unfavourite() { }
+
+}
