+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingArtShowMessage_EventDetailPresenterShould.swift
@@ -3,19 +3,6 @@ import EurofurenceModel
 import EurofurenceModelTestDoubles
 import XCTest
 
-struct StubArtShowEventViewModel: EventDetailViewModel {
-
-    var artShowMessageViewModel: EventArtShowMessageViewModel
-
-    var numberOfComponents: Int { return 1 }
-    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
-    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(artShowMessageViewModel
-        ) }
-    func favourite() { }
-    func unfavourite() { }
-
-}
-
 class WhenBindingArtShowMessage_EventDetailPresenterShould: XCTestCase {
 
     func testBindTheMessageOntoTheComponent() {
