+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingMainStageEvent_EventDetailPresenterShould.swift
@@ -3,19 +3,6 @@ import EurofurenceModel
 import EurofurenceModelTestDoubles
 import XCTest
 
-struct StubMainStageEventViewModel: EventDetailViewModel {
-
-    var mainStageMessageViewModel: EventMainStageMessageViewModel
-
-    var numberOfComponents: Int { return 1 }
-    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
-    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(mainStageMessageViewModel
-        ) }
-    func favourite() { }
-    func unfavourite() { }
-
-}
-
 class WhenBindingMainStageEvent_EventDetailPresenterShould: XCTestCase {
 
     func testBindTheMessageOntoTheComponent() {
