+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingPhotoshootBanner_EventDetailPresenterShould.swift
@@ -3,19 +3,6 @@ import EurofurenceModel
 import EurofurenceModelTestDoubles
 import XCTest
 
-struct StubPhotoshootEventViewModel: EventDetailViewModel {
-
-    var photoshootMessageViewModel: EventPhotoshootMessageViewModel
-
-    var numberOfComponents: Int { return 1 }
-    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
-    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(photoshootMessageViewModel
-        ) }
-    func favourite() { }
-    func unfavourite() { }
-
-}
-
 class WhenBindingPhotoshootBanner_EventDetailPresenterShould: XCTestCase {
 
     func testBindTheMessageOntoTheComponent() {
