+++ EurofurenceTests/Application/Components/Event Detail/Presenter Tests/WhenBindingSuperSponsorsOnlyMessage_EventDetailPresenterShould.swift
@@ -3,19 +3,6 @@ import EurofurenceModel
 import EurofurenceModelTestDoubles
 import XCTest
 
-struct StubSuperSponsorsOnlyEventViewModel: EventDetailViewModel {
-
-    var superSponsorsOnlyWarningViewModel: EventSuperSponsorsOnlyWarningViewModel
-
-    var numberOfComponents: Int { return 1 }
-    func setDelegate(_ delegate: EventDetailViewModelDelegate) { }
-    func describe(componentAt index: Int, to visitor: EventDetailViewModelVisitor) { visitor.visit(superSponsorsOnlyWarningViewModel
-        ) }
-    func favourite() { }
-    func unfavourite() { }
-
-}
-
 class WhenBindingSuperSponsorsOnlyMessage_EventDetailPresenterShould: XCTestCase {
 
     func testBindTheMessageOntoTheComponent() {
